Add HOME confirmation to init.sh

+++ init.sh
@@ -7,27 +7,33 @@
 
 
 # FILLME ----------------------------------------
-HOME="$HOME"
+HOMEC="$HOME"
 install_dotfiles=1
 install_packages=1
 # -----------------------------------------------
 
 
+# confirm home
+read -r -p "Is \$HOME supposed to be \"$HOMEC\"? [y/N] " response
+if ! [[ $response == "y" ]]; then
+  echo "Aborting..."
+  exit 0
+fi
+
 # define paths
-DEV="$HOME/dev"
+DEV="$HOMEC/dev"
 
 # exit on fail
 set -e
 
-
 # start setup
 if [ $install_dotfiles -eq 1 ]; then
     echo "# Install dotfiles..."
-    cd $HOME
+    cd $HOMEC
     git clone https://github.com/e9t/dotfiles.git
-    mv dotfiles/* dotfiles/.[^.]* $HOME
+    mv dotfiles/* dotfiles/.[^.]* $HOMEC
     rmdir dotfiles
-    sed -i "2s@.*@export HOME=\"$HOME\"@" $HOME/.bash_aliases
+    sed -i "2s@.*@export HOME=\"$HOMEC\"@" $HOMEC/.bash_aliases
     git submodule init
     git submodule update
 fi
@@ -41,9 +47,9 @@ if [ $install_packages -eq 1 ]; then
     git submodule init
     git submodule update
     git clone https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
-    cd $DEV/pkgs/fasd && PREFIX=$HOME make install
-    cd $DEV/pkgs/htop && ./autogen.sh && ./configure --prefix=$HOME && make && ln -s $PWD/htop $HOME/bin
-    cd $DEV/pkgs/tmux && ./autogen.sh && ./configure --exec_prefix="$HOME" && make && ln -s $PWD/tmux $HOME/bin
+    cd $DEV/pkgs/fasd && PREFIX=$HOMEC make install
+    cd $DEV/pkgs/htop && ./autogen.sh && ./configure --prefix=$HOMEC && make && ln -s $PWD/htop $HOMEC/bin
+    cd $DEV/pkgs/tmux && ./autogen.sh && ./configure --exec_prefix="$HOMEC" && make && ln -s $PWD/tmux $HOMEC/bin
 fi
 
 # vim:sw=4:ts=4:et
